server/requirements.txt
alembic==1.7.4
Babel==2.9.1
blinker==1.4
click==8.0.3
dnspython==2.1.0
email-validator==1.1.3
Flask==2.0.2
Flask-BabelEx==0.9.4
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-Principal==0.4.0
Flask-Security==3.0.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
greenlet==1.1.2
idna==3.3
importlib-metadata==4.8.1
importlib-resources==5.3.0
itsdangerous==2.0.1
Jinja2==3.0.2
Mako==1.1.5
MarkupSafe==2.0.1
passlib==1.7.4
psycopg2-binary==2.9.1
python-dotenv==0.19.1
pytz==2021.3
six==1.16.0
speaklater==1.3
SQLAlchemy==1.4.26
SQLAlchemy-Utils==0.37.9
Werkzeug==2.0.2
WTForms==3.0.0
zipp==3.6.0
Flask-Cors==3.0.10
